Copy ICC profile from the input to the output if requested
Use the number of channels specified in the input image, and use the
input colorspace (typically YCbCr), to avoid the need for conversion.
Then revert those two changes because they both make it slower.

Copy the ICC profile from the input image to the output image, if
requested in a command-line option.